max float at 1st position

+++ GenericTestMax.cs
@@ -6,11 +6,9 @@ using System.Threading.Tasks;
 
 namespace GenericsNew
 {
-
-
     public class GenericsMax
     {
-        public static int MaxIntNumber(int firstNum, int secondNum, int thirdNum)
+        public static double MaxIntNumber(double firstNum, double secondNum, double thirdNum)
         {
             if (firstNum.CompareTo(secondNum) > 0 && firstNum.CompareTo(thirdNum) > 0 || firstNum.CompareTo(secondNum) >= 0 && firstNum.CompareTo(thirdNum) > 0 || firstNum.CompareTo(secondNum) > 0 && firstNum.CompareTo(thirdNum) >= 0)
             {
@@ -29,4 +27,6 @@ namespace GenericsNew
             return firstNum;
         }
     }
+
+
 }
